Mother of refactor
