+++ lib/v2/domain-shared/page_state.dart
@@ -1,0 +1,2 @@
+/// --- STATES
+enum PageState { initial, loading, failure, success }
