+++ lib/v2/screens/profile/interactor/events.dart
@@ -1,0 +1,15 @@
+import 'package:equatable/equatable.dart';
+import 'package:flutter/material.dart';
+import 'package:meta/meta.dart';
+
+/// --- EVENTS
+@immutable
+abstract class ProfileEvent extends Equatable {
+  @override
+  List<Object> get props => [];
+}
+
+class LoadProfile extends ProfileEvent {
+  @override
+  String toString() => 'LoadProfile';
+}
